fixed ascii art alignment on alpha for good now
The issue was really the mixed tabs and spaces. I substituted each tab
for 4 spaces and it worked well.
What made it difficult to find out the issue was that once I zoomed out
at least once on the terminal the ascii art would get aligned. For some
reason the terminal's zoom (both wezterm and kitty) dissolves the
difference between tabs and spaces for as longs as it's zoomed out.

modified:   lua/user/alpha.lua